♻️ refactor(usage-service): split store repositories
Move store data shapes into internal/model and migrate SQLite table setup into repository/sqlite.

Add focused repositories for settings, usage events, dead letters, model prices, and API key aliases while keeping store.Store as the compatibility facade for existing callers.

The HTTP layer and collector still use the old store API, so this stage changes structure without changing API behavior or SQLite schema.
