+++ usage-service/internal/model/api_key_alias.go
@@ -1,0 +1,7 @@
+package model
+
+type APIKeyAlias struct {
+	APIKeyHash  string `json:"apiKeyHash"`
+	Alias       string `json:"alias"`
+	UpdatedAtMS int64  `json:"updatedAtMs"`
+}
