+++ usage-service/internal/model/model_price.go
@@ -1,0 +1,17 @@
+package model
+
+type ModelPrice struct {
+	Prompt        float64 `json:"prompt"`
+	Completion    float64 `json:"completion"`
+	Cache         float64 `json:"cache"`
+	Source        string  `json:"source,omitempty"`
+	SourceModelID string  `json:"sourceModelId,omitempty"`
+	RawJSON       string  `json:"rawJson,omitempty"`
+	UpdatedAtMS   int64   `json:"updatedAtMs,omitempty"`
+	SyncedAtMS    *int64  `json:"syncedAtMs,omitempty"`
+}
+
+type ModelPriceSyncResult struct {
+	Imported int `json:"imported"`
+	Skipped  int `json:"skipped"`
+}
