+++ usage-service/internal/model/setup.go
@@ -1,0 +1,14 @@
+package model
+
+import "fmt"
+
+type Setup struct {
+	CPAUpstreamURL string `json:"cpaBaseUrl"`
+	ManagementKey  string `json:"managementKey,omitempty"`
+	Queue          string `json:"queue,omitempty"`
+	PopSide        string `json:"popSide,omitempty"`
+}
+
+func (s Setup) String() string {
+	return fmt.Sprintf("upstream=%s queue=%s popSide=%s", s.CPAUpstreamURL, s.Queue, s.PopSide)
+}
