+++ usage-service/internal/repository/sqlite/options.go
@@ -1,0 +1,5 @@
+package sqlite
+
+type Options struct {
+	Path string
+}
