+++ usage-service/internal/repository/usageevent/repository.go
@@ -1,0 +1,234 @@
+package usageevent
+
+import (
+	"context"
+	"database/sql"
+	"encoding/json"
+
+	"github.com/seakee/cpa-manager/usage-service/internal/model"
+)
+
+type Repository interface {
+	InsertBatch(ctx context.Context, events []model.UsageEvent) (model.InsertResult, error)
+	ListRecent(ctx context.Context, limit int) ([]model.UsageEvent, error)
+	Count(ctx context.Context) (int64, error)
+	ExportJSONL(ctx context.Context) ([]byte, error)
+}
+
+type repository struct {
+	db *sql.DB
+}
+
+func New(db *sql.DB) Repository {
+	return &repository{db: db}
+}
+
+func (r *repository) InsertBatch(ctx context.Context, events []model.UsageEvent) (model.InsertResult, error) {
+	if len(events) == 0 {
+		return model.InsertResult{}, nil
+	}
+	tx, err := r.db.BeginTx(ctx, nil)
+	if err != nil {
+		return model.InsertResult{}, err
+	}
+	defer func() {
+		_ = tx.Rollback()
+	}()
+
+	stmt, err := tx.PrepareContext(ctx, `insert or ignore into usage_events (
+		request_id, event_hash, timestamp_ms, timestamp, provider, model, endpoint, method, path,
+		auth_type, auth_index, source, source_hash, api_key_hash,
+		account_snapshot, auth_label_snapshot, auth_file_snapshot, auth_provider_snapshot, auth_snapshot_at_ms,
+		input_tokens, output_tokens, reasoning_tokens, cached_tokens, cache_tokens, total_tokens,
+		latency_ms, failed, raw_json, created_at_ms
+	) values (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`)
+	if err != nil {
+		return model.InsertResult{}, err
+	}
+	defer stmt.Close()
+
+	result := model.InsertResult{}
+	for _, event := range events {
+		failed := 0
+		if event.Failed {
+			failed = 1
+		}
+		res, err := stmt.ExecContext(
+			ctx,
+			nullString(event.RequestID),
+			event.EventHash,
+			event.TimestampMS,
+			event.Timestamp,
+			nullString(event.Provider),
+			event.Model,
+			nullString(event.Endpoint),
+			nullString(event.Method),
+			nullString(event.Path),
+			nullString(event.AuthType),
+			nullString(event.AuthIndex),
+			nullString(event.Source),
+			nullString(event.SourceHash),
+			nullString(event.APIKeyHash),
+			nullString(event.AccountSnapshot),
+			nullString(event.AuthLabelSnapshot),
+			nullString(event.AuthFileSnapshot),
+			nullString(event.AuthProviderSnapshot),
+			nullPositiveInt64(event.AuthSnapshotAtMS),
+			event.InputTokens,
+			event.OutputTokens,
+			event.ReasoningTokens,
+			event.CachedTokens,
+			event.CacheTokens,
+			event.TotalTokens,
+			nullInt(event.LatencyMS),
+			failed,
+			nullString(event.RawJSON),
+			event.CreatedAtMS,
+		)
+		if err != nil {
+			return model.InsertResult{}, err
+		}
+		affected, _ := res.RowsAffected()
+		if affected > 0 {
+			result.Inserted++
+		} else {
+			result.Skipped++
+		}
+	}
+	if err := tx.Commit(); err != nil {
+		return model.InsertResult{}, err
+	}
+	return result, nil
+}
+
+func (r *repository) ListRecent(ctx context.Context, limit int) ([]model.UsageEvent, error) {
+	if limit <= 0 {
+		limit = 50000
+	}
+	rows, err := r.db.QueryContext(ctx, `select
+		request_id, event_hash, timestamp_ms, timestamp, provider, model, endpoint, method, path,
+		auth_type, auth_index, source, source_hash, api_key_hash,
+		account_snapshot, auth_label_snapshot, auth_file_snapshot, auth_provider_snapshot, auth_snapshot_at_ms,
+		input_tokens, output_tokens, reasoning_tokens, cached_tokens, cache_tokens, total_tokens,
+		latency_ms, failed, raw_json, created_at_ms
+		from usage_events
+		order by timestamp_ms desc, id desc
+		limit ?`, limit)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+
+	events := make([]model.UsageEvent, 0)
+	for rows.Next() {
+		var event model.UsageEvent
+		var requestID, provider, endpoint, method, path, authType, authIndex, source, sourceHash, apiKeyHash, accountSnapshot, authLabelSnapshot, authFileSnapshot, authProviderSnapshot, rawJSON sql.NullString
+		var authSnapshotAt sql.NullInt64
+		var latency sql.NullInt64
+		var failed int
+		if err := rows.Scan(
+			&requestID,
+			&event.EventHash,
+			&event.TimestampMS,
+			&event.Timestamp,
+			&provider,
+			&event.Model,
+			&endpoint,
+			&method,
+			&path,
+			&authType,
+			&authIndex,
+			&source,
+			&sourceHash,
+			&apiKeyHash,
+			&accountSnapshot,
+			&authLabelSnapshot,
+			&authFileSnapshot,
+			&authProviderSnapshot,
+			&authSnapshotAt,
+			&event.InputTokens,
+			&event.OutputTokens,
+			&event.ReasoningTokens,
+			&event.CachedTokens,
+			&event.CacheTokens,
+			&event.TotalTokens,
+			&latency,
+			&failed,
+			&rawJSON,
+			&event.CreatedAtMS,
+		); err != nil {
+			return nil, err
+		}
+		event.RequestID = requestID.String
+		event.Provider = provider.String
+		event.Endpoint = endpoint.String
+		event.Method = method.String
+		event.Path = path.String
+		event.AuthType = authType.String
+		event.AuthIndex = authIndex.String
+		event.Source = source.String
+		event.SourceHash = sourceHash.String
+		event.APIKeyHash = apiKeyHash.String
+		event.AccountSnapshot = accountSnapshot.String
+		event.AuthLabelSnapshot = authLabelSnapshot.String
+		event.AuthFileSnapshot = authFileSnapshot.String
+		event.AuthProviderSnapshot = authProviderSnapshot.String
+		if authSnapshotAt.Valid {
+			event.AuthSnapshotAtMS = authSnapshotAt.Int64
+		}
+		event.RawJSON = rawJSON.String
+		event.Failed = failed != 0
+		if latency.Valid {
+			value := latency.Int64
+			event.LatencyMS = &value
+		}
+		events = append(events, event)
+	}
+	return events, rows.Err()
+}
+
+func (r *repository) Count(ctx context.Context) (int64, error) {
+	var count int64
+	if err := r.db.QueryRowContext(ctx, `select count(*) from usage_events`).Scan(&count); err != nil {
+		return 0, err
+	}
+	return count, nil
+}
+
+func (r *repository) ExportJSONL(ctx context.Context) ([]byte, error) {
+	events, err := r.ListRecent(ctx, 0)
+	if err != nil {
+		return nil, err
+	}
+	output := make([]byte, 0)
+	for i := len(events) - 1; i >= 0; i-- {
+		line, err := json.Marshal(events[i])
+		if err != nil {
+			return nil, err
+		}
+		output = append(output, line...)
+		output = append(output, '\n')
+	}
+	return output, nil
+}
+
+func nullString(value string) any {
+	if value == "" {
+		return nil
+	}
+	return value
+}
+
+func nullInt(value *int64) any {
+	if value == nil {
+		return nil
+	}
+	return *value
+}
+
+func nullPositiveInt64(value int64) any {
+	if value <= 0 {
+		return nil
+	}
+	return value
+}
